Перенес hasDuplicates в utils и обработчики в img-upload
