+++ js/text-validation.js
@@ -1,12 +1,12 @@
+import {hasDuplicates} from './utils.js';
+
 const uploadForm = document.querySelector('.img-upload__form');
 const tagsInput = uploadForm.querySelector('.text__hashtags');
+const description = uploadForm.querySelector('.text__description');
+const pattern = /^#(?=.*[^0-9])[a-zа-яё0-9]{1,19}$/i;
 const MAX_TAGS_NUMBER = 5;
 const MAX_TAGS_LENGTH = 20;
-const pattern = /^#(?=.*[^0-9])[a-zа-яё0-9]{1,19}$/i;
-const description = uploadForm.querySelector('.text__description');
 const MAX_DESCRIPTION_LENGTH = 140;
-
-tagsInput.addEventListener('input', onTagsInputInput);
 
 function onTagsInputInput () {
   const tags = tagsInput.value.split(' ');
@@ -29,20 +29,6 @@ function onTagsInputInput () {
   }
 }
 
-function hasDuplicates (array) {
-  let valuesSoFar = Object.create(null);
-  for (let i = 0; i < array.length; ++i) {
-    let value = array[i].toLowerCase();
-    if (value in valuesSoFar) {
-      return true;
-    }
-    valuesSoFar[value] = true;
-  }
-  return false;
-}
-
-description.addEventListener('input', onDescriptionInput);
-
 function onDescriptionInput () {
   if (description.value.length > MAX_DESCRIPTION_LENGTH) {
     description.setCustomValidity('Максимум ' + MAX_DESCRIPTION_LENGTH + ' знаков. Сделайте короче на ' + (description.value.length - MAX_DESCRIPTION_LENGTH) + '.');
@@ -50,3 +36,5 @@ function onDescriptionInput () {
     description.setCustomValidity('');
   }
 }
+
+export {onTagsInputInput, onDescriptionInput}
